Towers and Placment
I've fully added the Towers Hud, 2 Towers, and you can place towers in the game. I also made it so that if 3 enemies finish their path it will print game over. I also make it so that if you select a tower and want to delete it while holding it, you can just place it where the hub is and it will delete it.